// src/api/generated/schemas/duckReactionResponse.ts
/**
 * Generated by orval v7.16.1 🍺
 * Do not edit manually.
 * Duck Party API
 * API for Duck Party backend service
 * OpenAPI spec version: 1.0
 */

import type { DuckReactionResponseReaction } from "./duckReactionResponseReaction";
import type { DuckResponse } from "./duckResponse";
import type { DuckUserResponse } from "./duckUserResponse";

export interface DuckReactionResponse {
  created_at?: string;
  duck?: DuckResponse;
  duck_id?: number;
  reaction?: DuckReactionResponseReaction;
  user?: DuckUserResponse;
  user_id?: number;
}